fix(store): recreate checkout instead of hide/unhide
Old way of hiding and unhinding checkout overlay caused a shit ton
of bugs and headakes. This solves issues in regards to getting
phone nubmer, being charged multiple times because event listeners
would've been added multiple times. This however is not a perfect
solution because charge() is now doing cleanup that should have
been done in checkout(). createPaymentMethod should probably be
called in checkout() and return the payment method with customer.