registration: route new members to Spond group
After successful registration members are transfered to spond group
where they can request membership. Previously there were no obvious way
for users to know that they need to join spond to get access to join
practices. Url is hardcoded in this case which is not optimal but the
link does not need to change between years so this "temporary" solution
should not break in the near future. An idea is also to have some sort
of CI rule to check that the link is returning a 200 response.