translator: throw on missing translations
This patch fixes some bugs with obsessive error logging when it cannot
find translations for pages that does not exists. In stead check if page
in queston exists first before trying to access it. This will lead to
less useless logging and we can throw the errors in stead of logging
them. This patch also includes logging of requests that "should" return
404. This is useful for spotting bad links and also for detecting
     intrution attempts.